# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TotalEasy API
  contact:
    email: marcosvisentini@gmail.com
  license:
    name: GPLv3
    url: https://www.gnu.org/licenses/gpl-3.0.html
  version: v1
servers:
  - url: http://localhost/api
    description: Desenvolvimento
...
paths:
...
  /boletins-urna:
    get:
      tags:
        - boletins-urna
      summary: 'Encontra um ou mais boletins de urna.'
      description: 'Se o parâmetro `id` for informado, é retornado o boletim de urna identificado por ele. Se não, são retornados todos os boletins de urna.'
      operationId: findBoletinsUrna
      parameters:
        - in: query
          name: id
          description: 'Os dados que identificam o boletim de urna.'
          schema:
            $ref: '#/components/schemas/BoletimUrnaIdDTO'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/BoletimUrnaRetrievalDTOArray'
                  - $ref: '#/components/schemas/BoletimUrnaRetrievalDTO'
        '4XX':
          description: 'Client error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
        - bearerAuth: []
    post:
      tags:
        - boletins-urna
      summary: 'Constrói um boletim de urna a partir de um ou mais payloads de QR code.'
      description: 'Constrói um boletim de urna a partir de um ou mais payloads de QR code.'
      operationId: buildBoletimUrna
      requestBody:
        description: 'O(s) payload(s) do(s) QR code(s) do boletim de urna.'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoletimUrnaBuildDTO'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '4XX':
          description: 'Client error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
        - bearerAuth: []
...
  /urnas-eletronicas:
    get:
      tags:
        - urnas-eletronicas
      summary: 'Encontra todas as urnas eletrônicas.'
      description: 'Encontra todas as urnas eletrônicas.'
      operationId: 'findUrnasEletronicas'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UrnaEletronicaRetrievalDTO'
        '4XX':
          description: 'Client error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
        - bearerAuth: []
  /urnas-eletronicas/{numeroSerie}:
    get:
      tags:
        - urnas-eletronicas
      summary: 'Encontra uma urna eletrônica.'
      description: 'Encontra uma urna eletrônica.'
      operationId: findUrnaEletronica
      parameters:
        - name: numeroSerie
          in: path
          description: 'O número de série da urna eletrônica.'
          required: true
          schema:
            type: integer
            format: int64
            example: 1678798
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UrnaEletronicaRetrievalDTO'
        '4XX':
          description: 'Client error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
        - bearerAuth: []
  /urnas-eletronicas/{numeroSerie}/boletins-urna:
    get:
      tags:
        - urnas-eletronicas
      summary: 'Encontra todos os boletins de urna de uma urna eletrônica.'
      description: 'Encontra todos os boletins de urna de uma urna eletrônica.'
      operationId: findBoletinsUrnaUrnaEletronica
      parameters:
        - name: numeroSerie
          in: path
          description: 'O número de série da urna eletrônica.'
          required: true
          schema:
            type: integer
            format: int64
            example: 1678798
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BoletimUrnaRetrievalDTO'
        '4XX':
          description: 'Client error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    RegiaoRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 6
        codigoIBGE:
          type: integer
          format: int64
          example: 4
        nome:
          type: string
          example: Sul
        sigla:
          type: string
          example: S
    UFRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 22
        codigoIBGE:
          type: integer
          format: int64
          example: 43
        nome:
          type: string
          example: 'Rio Grande do Sul'
        sigla:
          type: string
          example: RS
        regiao:
          $ref: '#/components/schemas/RegiaoRetrievalDTO'
    ZonaRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 34
        numeroTSE:
          type: integer
          format: int64
          example: 18
        uf:
          $ref: '#/components/schemas/UFRetrievalDTO'
    SecaoRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 61
        numeroTSE:
          type: integer
          format: int64
          example: 112
        zona:
          $ref: '#/components/schemas/ZonaRetrievalDTO'
    OrigemConfiguracaoProcessoEleitoralRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nome:
          type: string
          example: 'Eleição legal oficial'
        nomeAbreviado:
          type: string
          example: LEG
    ProcessoEleitoralRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        codigoTSE:
          type: integer
          format: int64
          example: 395
        nome:
          type: string
          example: 'Eleições Gerais 2022'
        origemConfiguracao:
          $ref: '#/components/schemas/OrigemConfiguracaoProcessoEleitoralRetrievalDTO'
    PleitoRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        codigoTSE:
          type: integer
          format: int64
          example: 406
        nome:
          type: string
          example: '1º turno'
        turno:
          type: integer
          format: int64
          example: 1
        data:
          type: string
          format: date
          example: 02/10/2022
        processoEleitoral:
          $ref: '#/components/schemas/ProcessoEleitoralRetrievalDTO'
    SecaoPleitoRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 61
        secao:
          $ref: '#/components/schemas/SecaoRetrievalDTO'
        pleito:
          $ref: '#/components/schemas/PleitoRetrievalDTO'
    FaseRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        codigoTSE:
          type: integer
          format: int64
          example: 2
        nome:
          type: string
          example: Oficial
...
    OrigemBoletimUrnaRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nome:
          type: string
          example: 'Software de Votação'
        nomeAbreviado:
          type: string
          example: Vota
    UrnaEletronicaRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 61
        numeroSerie:
          type: integer
          format: int64
          example: 1303045
        codigoIdentificacaoCarga:
          type: string
          example: '904585283618888740610421'
        versaoSoftware:
          type: string
          example: '8.26.0.0'
        dataAbertura:
          type: string
          format: date
          example: 02/10/2022
        horarioAbertura:
          type: string
          format: time
          example: 08:00:01
        dataFechamento:
          type: string
          format: date
          example: 02/10/2022
        horarioFechamento:
          type: string
          format: time
          example: 17:04:48
    BoletimUrnaRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 61
        secaoPleito:
          $ref: '#/components/schemas/SecaoPleitoRetrievalDTO'
        fase:
          $ref: '#/components/schemas/FaseRetrievalDTO'
        origem:
          $ref: '#/components/schemas/OrigemBoletimUrnaRetrievalDTO'
        urnaEletronica:
          $ref: '#/components/schemas/UrnaEletronicaRetrievalDTO'
        assinatura:
          type: string
          example: A68CF51ECA1A79BB80D5514012AB8D2E7424291023FB42E0CFEB4849D7337720434E8D30C798230B5695D1ADF19B2142AAC1687A1C6D81BC37A406041E238D00
        quantidadeTotalQRCodes:
          type: integer
          format: int64
          example: 3
        dataEmissao:
          type: string
          format: date
          example: 02/10/2022
        horarioEmissao:
          type: string
          format: time
          example: 17:05:53
...
    BoletimUrnaBuildDTO:
      type: object
      required:
        - payloads
      properties:
        payloads:
          type: array
          items:
            type: string
          example: ['QRBU:1:3 VRQR:1.5 VRCH:20220829 ORIG:VOTA ORLC:LEG PROC:395 DTPL:20221002 PLEI:406 TURN:1 FASE:O UNFE:RS MUNI:84891 ZONA:11 SECA:129 IDUE:1678798 IDCA:614979838198585678607730 VERS:8.26.0.0 LOCA:1112 APTO:327 COMP:266 FALT:61 HBMA:30 DTAB:20221002 HRAB:080001 DTFC:20221002 HRFC:171020 IDEL:546 CARG:6 TIPO:1 VERC:202209131721 PART:10 1000:1 1010:2 1014:1 1022:3 1040:1 1044:1 LEGP:0 TOTP:9 PART:11 1111:2 1144:111 1155:1 1166:3 1199:1 LEGP:2 TOTP:120 PART:12 1212:4 1234:2 LEGP:0 TOTP:6 PART:13 1307:5 1355:49 1370:3 LEGP:4 TOTP:61 PART:14 1414:1 LEGP:0 TOTP:1 PART:15 1500:1 1501:1 1515:4 1590:2 LEGP:0 TOTP:8 PART:19 1901:1 1909:1 1922:2 1930:1 1944:1 1977:4 LEGP:0 TOTP:10 PART:22 2200:2 2202:1 2210:1 2214:1 2234:1 2244:1 LEGP:0 TOTP:7 PART:23 2323:3 LEGP:0 TOTP:3 PART:30 3000:1 3030:5 LEGP:0 TOTP:6 PART:43 4343:2 LEGP:0 TOTP:2 PART:44 4400:1 4312AFFAC5FFE88683F14A1FF09A79E9D71E084834A3713F43ABF4592BF7C861187DD49D65791442AF3CC6D88DB187BE7860237F5F29CCEFC5069637C328885A', 'QRBU:2:3 VRQR:1.5 VRCH:20220829 4411:1 4422:1 LEGP:0 TOTP:3 PART:45 4511:2 4545:1 4566:1 LEGP:0 TOTP:4 PART:50 5010:1 5050:2 LEGP:0 TOTP:3 PART:55 5512:1 5577:2 LEGP:0 TOTP:3 PART:77 7797:1 LEGP:0 TOTP:1 APTA:327 NOMI:241 LEGC:6 BRAN:13 NULO:6 TOTC:266 CARG:7 TIPO:1 VERC:202209131721 PART:10 10022:2 10040:2 10122:1 10200:1 10222:3 10300:2 10800:4 LEGP:1 TOTP:16 PART:11 11000:1 11011:1 11012:79 11111:2 11190:1 11240:6 11411:1 11555:29 11777:1 11999:3 LEGP:4 TOTP:128 PART:12 12050:1 12220:3 12777:3 LEGP:1 TOTP:8 PART:13 13007:1 13114:8 13120:2 13456:10 13655:3 13690:2 13813:12 13888:2 LEGP:9 TOTP:49 PART:15 15015:1 15200:3 15615:1 15777:1 15789:1 LEGP:0 TOTP:7 PART:16 LEGP:1 TOTP:1 PART:19 19140:2 19177:2 19234:1 19300:1 19444:2 19777:1 LEGP:1 TOTP:10 PART:20 20222:1 LEGP:0 TOTP:1 PART:22 22000:1 22212:1 22214:1 22222:1 LEGP:1 TOTP:5 PART:23 25C826B103C37BE389F9054AF7307A68E75F195C0E7EF2C3D93152E956C1E61EC75CF6C8C3C69D41E18E66CB9AF9E4DAA63B04B8FE4965717CB85D45AC455F79', 'QRBU:3:3 VRQR:1.5 VRCH:20220829 23456:1 LEGP:0 TOTP:1 PART:28 28000:2 LEGP:0 TOTP:2 PART:30 30456:1 LEGP:0 TOTP:1 PART:44 44000:1 44190:1 44192:2 44560:1 LEGP:0 TOTP:5 PART:45 45000:1 45111:1 45200:1 LEGP:0 TOTP:3 PART:50 50000:2 50123:1 LEGP:0 TOTP:3 PART:55 55123:1 LEGP:1 TOTP:2 PART:77 77377:1 LEGP:0 TOTP:1 APTA:327 NOMI:224 LEGC:19 BRAN:15 NULO:8 TOTC:266 CARG:5 TIPO:0 VERC:202209131721 100:78 111:24 131:101 200:1 555:22 700:1 APTA:327 NOMI:227 BRAN:20 NULO:19 TOTC:266 CARG:3 TIPO:0 VERC:202209131721 11:67 12:2 13:82 20:5 22:57 30:1 45:33 APTA:327 NOMI:247 BRAN:10 NULO:9 TOTC:266 IDEL:544 CARG:1 TIPO:0 VERC:202209081640 12:4 13:136 14:1 15:6 22:105 30:1 44:1 APTA:327 NOMI:254 BRAN:4 NULO:8 TOTC:266 8E000C30521E58CF41F2D1E86FA6198DD1AAEFCE1AAA3B03FAF6F0791E45A46F70BAB73D4A73CF8182229A0E8048262001857BAF532207FE30514C857A0D3EF6 0B61D74201F308D7899961BD4D58B7AAE5F436C287AC3DDC4AA312DEAC14EC24FD8F30F0BF140FE22F5B9D2913D417359C6C0A2A3D7C096EA4FE1284ED092504']
...
    BoletimUrnaIdDTO:
      type: object
      required:
        - numeroTSESecao
        - numeroTSEZona
        - siglaUF
        - codigoTSEPleito
      properties:
        numeroTSESecao:
          type: integer
          format: int64
          example: 112
        numeroTSEZona:
          type: integer
          format: int64
          example: 18
        siglaUF:
          type: string
          minLength: 2
          maxLength: 2
          example: RS
        codigoTSEPleito:
          type: integer
          format: int64
          example: 406
...
    ApiResponse:
      type: object
      required:
        - status
        - content
        - timestamp
      properties:
        status:
          type: string
          example: OK
        content:
          type: string
          example: 'Conteúdo da resposta'
        timestamp:
          type: string
          format: date-time
          example: '03/03/2023 16:25:00'
    ApiError:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: 'Descrição detalhada de um erro'
    ApiErrorResponse:
      type: object
      required:
        - status
        - errors
        - timestamp
      properties:
        status:
          type: string
          example: BAD_REQUEST
        description:
          type: string
          example: 'Descrição genérica de um ou mais erros'
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ApiError'
        timestamp:
          type: string
          format: date-time
          example: '05/03/2023 09:43:11'
...
    BoletimUrnaRetrievalDTOArray:
        type: array
        items:
          $ref: '#/components/schemas/BoletimUrnaRetrievalDTO'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT