# src/main/java/app/totaleasy/backend/rest/dto/retrieval/SecaoRetrievalDTO.java
package app.totaleasy.backend.rest.dto.retrieval;

import com.fasterxml.jackson.annotation.JsonPropertyOrder;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@JsonPropertyOrder(value = {"id", "numeroTSE", "zona"})
public class SecaoRetrievalDTO {

    private Integer id;

    private Integer numeroTSE;

    private ZonaRetrievalDTO zona;
}